% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{appeals_dict}
\alias{appeals_dict}
\title{Appeal reason codes used by the CCAO AS/400 system}
\format{
A data frame with 78 rows and 2 variables:
\describe{
  \item{reason_code}{Numeric code used in the AS/400 system}
  \item{reason_desc}{Text description/translation of the numeric code}
}
}
\usage{
appeals_dict
}
\description{
A dataset containing a lookup of codes for appeal rejection/acceptance. These
codes are primarily used in the DKEMPL SQL table. See the \href{https://prodassets.cookcountyassessor.com/s3fs-public/form_documents/reasoncodes.pdf}{CCAO website} for a PDF version.
}
\keyword{datasets}
