% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccao_colors}
\alias{ccao_colors}
\title{Official CCAO color palette}
\format{
A named list containing 12 CCAO colors
}
\usage{
ccao_colors
}
\description{
The CCAO Communications Department created a palette of colors used widely
throughout CCAO press materials and visualizations. Navy, gold, and
buttermilk are the colors used in the CCAO logo.
}
\keyword{datasets}
