% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccao_funs.R
\name{ccao_funs}
\alias{ccao_funs}
\alias{ccao_cod}
\alias{ccao_prd}
\alias{ccao_prb}
\title{Calculate sales ratio statistics using CCAO SOPs}
\usage{
ccao_cod(ratio, suppress = FALSE, na.rm = FALSE)

ccao_prd(assessed, sale_price, suppress = FALSE, na.rm = FALSE)

ccao_prb(assessed, sale_price, suppress = FALSE, na.rm = FALSE)
}
\arguments{
\item{ratio}{A numeric vector of ratios centered around 1, where the
numerator of the ratio is the estimated fair market value and the
denominator is the actual sale price. NOTE: CCAO typically uses lagged
or leading ratios to lessen the effect of sales chasing.}

\item{suppress}{Default FALSE. A boolean value indicating whether or not
to ignore minimum sample size requirements (N must be >= 30). If TRUE when
N < 30, will return NAs.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}

\item{assessed}{A numeric vector of assessed values. Must be the same
length as \code{sale_price}.}

\item{sale_price}{A numeric vector of sale prices. Must be the same length
as \code{assessed}.}
}
\value{
A named list containing the statistic, its 95\% confidence interval,
whether or not the statistic meets the IAAO standard, whether or not the
95\% confidence interval overlaps the IAAO standard, and the number of
observations used to calculate the statistic (after outliers are removed).
}
\description{
Calculate sales ratio statistics (COD, PRD, PRB) according
to Cook County Assessor's Office Standard Operating Procedures (SOPs).

The main SOPs affecting these calculations are:
\enumerate{
\item Outlier trimming. Sales ratio performance statistics are quite sensitive
to outliers. As a result, the top and bottom 5\\% of the input vector(s) are
always dropped.
\item Minimum sample size. Sales ratio performance statistics can be inaccurate
or misleading when the input sample is small. As a result, samples with less
than 30 observations after trimming outliers will stop execution or will
return NA if \code{suppress = TRUE}.
}

For more information on how each statistic or its confidence interval is
calculated. See its respective function in the
\href{https://gitlab.com/ccao-data-science---modeling/packages/assessr}{AssessR package}.
}
\section{Functions}{
\itemize{
\item \code{ccao_cod}: Return named list of CCAO SOP compliant COD statistics.

\item \code{ccao_prd}: Return named list of CCAO SOP compliant PRD statistics.

\item \code{ccao_prb}: Return named list of CCAO SOP compliant PRB statistics.
}}

\examples{

# Load sample data from assessr
df <- assessr::ratios_sample

# Calculate performance stats using CCAO SOP defaults
ccao_cod(df$ratio)
ccao_prd(df$assessed, df$sale_price)
ccao_prb(df$assessed, df$sale_price)
}
\concept{ccao_funs}
