% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdu_dict}
\alias{cdu_dict}
\title{Codes used by the CCAO to identify certain distinct property situations}
\format{
A data frame with 47 rows and 4 variables:
\describe{
  \item{cdu_code}{Two-letter code containing the CDU, as shown in the AS/400}
  \item{cdu_type}{Class/type of property the CDU applies to}
  \item{cdu_desc}{Full description of the CDU}
  \item{cdu_desc_short}{Short description of the CDU}
}
}
\usage{
cdu_dict
}
\description{
A dataset containing a lookup of CDU codes. These codes are kind of a mess
and have been created and used inconsistently over the years.
}
\details{
They can indicate special property types or specific incentives applied to
a property.
}
\keyword{datasets}
