% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars_funs.R
\name{chars_288_active}
\alias{chars_288_active}
\title{Return active years given a 288 start date and township}
\usage{
chars_288_active(start_year, town)
}
\arguments{
\item{start_year}{A numeric vector of start years. Must be either 1 long or
the same length as \code{town}.}

\item{town}{A character vector of town codes or names. Must be either 1 long
or the same length as \code{start_year}.}
}
\value{
A list of numeric vectors containing the active years for each value
  of the input vectors. The list will always be N long, where N is the
  length of the longest input vector.
}
\description{
The State of Illinois has a home improvement exemption program
which allows property owners to deduct up to $75,000 per year of any value
created by improvements to a residential property.

This has the effect of essentially "freezing" a home's characteristics at
whatever they were prior to the start of the improvement project. For
example, if a property owner adds an additional bedroom and applies for a
288, the property will be valued as if the new bedroom does not exist until
the 288 expires and as long as the increase in valuation is less than
$75,000.

\href{https://www.ilga.gov/legislation/ilcs/fulltext.asp?DocName=003502000K15-180}{Per Illinois statute}, # nolint
288s expire after 4 years or until the next assessment cycle, whichever is
longer. For example, a 288 received in 2016 for a property in Northfield
(with assessment years 2016, 2019, and 2022) will last 6 years (until 2021,
the year before the 2022 reassessment).

This function calculates the years a 288 will be active given a start year
and the township of a property.
}
\examples{

# Simplest case to get the length for one property
chars_288_active(2016, "Northfield")

# Multiple years for a single township
chars_288_active(c(2016, 2010, 2020), "77")
chars_288_active(c(2016, 2010, 2020), "Evanston")

# Multiple townships for a single year
chars_288_active(2015, c("77", "Evanston", "10"))

# Pairwise lookup (2015 for Evanston, 2017 for Northfield)
chars_288_active(c(2015, 2017), c("Evanston", "Northfield"))
}
\seealso{
Other chars_funs: 
\code{\link{chars_fix_age}()},
\code{\link{chars_sparsify}()},
\code{\link{chars_update}()}
}
\concept{chars_funs}
