% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chars_cols}
\alias{chars_cols}
\title{List crosswalk of Athena columns and their ADDCHARS equivalents}
\format{
A nested list with 2 items:
\describe{
  \item{add}{Source and target columns with additive characteristics}
  \item{replace}{Source and target columns with characteristics to be
  replaced}
}
}
\usage{
chars_cols
}
\description{
A nested list containing vectors of column names that match those found in
the iasWorld DWELDAT and legacy ADDCHARS SQL tables. Can be used to translate
between tables.
}
\keyword{datasets}
