% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars_funs.R
\name{chars_fix_age}
\alias{chars_fix_age}
\title{Fix the AGE variable in CCAO data}
\usage{
chars_fix_age(age, year, town)
}
\arguments{
\item{age}{A numeric vector of ages. Must be either 1 long or the same length
as one of the other two inputs.}

\item{year}{A numeric vector of tax years. Usually the TAX_YEAR column.
Must be either 1 long or the same length as one of the other two inputs.}

\item{town}{A character vector of town codes or names. Must be either 1 long
or the same length as one of the other two inputs.}
}
\value{
A numeric vector of "true" ages the same length as the longest input
  vector.
}
\description{
The AGE variable in many CCAO datasets only updates when a
property is reassessed. This function will calculate the correct age of
a property given a township, current year, and age. It can be used with
\code{mutate()} to calculate a "true" age column.
}
\examples{

# Simplest case, get true age of one property
chars_fix_age(80, 2015, "Evanston")

# Getting many ages for different towns
chars_fix_age(80, 2015, c("Evanston", "Niles"))

# Creating mock data then fixing a column
df <- dplyr::tibble(
  age = c(120, 120, 123),
  year = c(2014, 2015, 2016),
  town = rep("25", 3)
)

df$true_age <- chars_fix_age(df$age, df$year, df$town)
}
\seealso{
Other chars_funs: 
\code{\link{chars_288_active}()},
\code{\link{chars_sparsify}()},
\code{\link{chars_update}()}
}
\concept{chars_funs}
