% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chars_sample_addchars}
\alias{chars_sample_addchars}
\title{Sample dataset from the ADDCHARS SQL table}
\format{
An object of class \code{data.frame} with 10 rows and 62 columns.
}
\source{
This data was extracted from SQL manually on 2021-01-20.
}
\usage{
chars_sample_addchars
}
\description{
A dataset containing a small subsample of rows from the ADDCHARS table. This
sample can be used with \code{chars_sparsify()} to generate sparse data
frames suitable for joining onto \code{chars_sample_universe}.
}
\keyword{datasets}
