% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chars_sample_hie}
\alias{chars_sample_hie}
\title{Sample dataset from the ccao.hie Athena SQL table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11 rows and 35 columns.
}
\source{
This data was extracted from Athena manually on 2022-03-08.
}
\usage{
chars_sample_hie
}
\description{
A dataset containing a small subsample of rows from the ccao.hie table,
which is a cleaned up version of the AS/400 ADDCHARS table. Matches the PINs
in the \code{\link{chars_sample_athena}} dataset.
}
\keyword{datasets}
