% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chars_sample_universe}
\alias{chars_sample_universe}
\title{Sample dataset from the VW_RES_UNIVERSE SQL view}
\format{
An object of class \code{data.frame} with 20 rows and 61 columns.
}
\source{
This data was extracted from SQL manually on 2021-01-20.
}
\usage{
chars_sample_universe
}
\description{
A dataset containing a small subsample of rows from the VW_RES_UNIVERSE view.
}
\keyword{datasets}
