% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chars_sample_universe_updated}
\alias{chars_sample_universe_updated}
\title{Sample dataset from the VW_RES_UNIVERSE SQL view with 288s applied}
\format{
An object of class \code{data.frame} with 20 rows and 62 columns.
}
\source{
Hand-updated and validated version of \code{chars_sample_universe}
with 288s applied.
}
\usage{
chars_sample_universe_updated
}
\description{
A dataset containing the properties in \code{chars_sample_universe}, but
updated with \code{chars_sparsity()} and \code{chars_update()}. This data is
used as a benchmark to unit test the 288-related functions. It was manually
validated against properties updated by the AS/400.
}
\keyword{datasets}
