% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars_funs.R
\name{chars_update}
\alias{chars_update}
\title{Update characteristic values using joined values from ADDCHARS}
\usage{
chars_update(data, additive_target, replacement_target)
}
\arguments{
\item{data}{A data frame containing CCAOSFCHARS columns AND their equivalent
ADDCHARS columns.}

\item{additive_target}{A tidyselect selection of CCAOSFCHARS column names
that should have characteristics added to them from their equivalent
ADDCHARS column, which must also be present.}

\item{replacement_target}{A tidyselect selection of CCAOSFCHARS column names
that should have characteristics replaced using their equivalent ADDCHARS
column, which must also be present.}
}
\value{
A data frame with updated additive and replacement target columns.
}
\description{
Function used to update the characteristic values of a
data frame containing both the original characteristic value (CCAOSFCHARS)
and the value to add or replace it with (ADDCHARS). This function expect a
data frame formatted using \code{\link{chars_sparsify}}.
}
\seealso{
Other chars_funs: 
\code{\link{chars_288_active}()},
\code{\link{chars_fix_age}()},
\code{\link{chars_sparsify}()}
}
\concept{chars_funs}
