% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coe_codes}
\alias{coe_codes}
\title{Data frame of Certificate of Error reason codes}
\format{
A data frame with 95 rows and 2 variables:
\describe{
  \item{coe_code}{The numeric code for a given CoE}
  \item{coe_reason}{The text explaination for the CoE numeric code}
}
}
\usage{
coe_codes
}
\description{
A dataset containing numeric codes and corresponding text explanations for
Certificates of Error (CoEs). These are issued when the Assessor's office
makes some kind of mistake that leads to an erroneous valuation.
}
\keyword{datasets}
