% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_funs.R
\name{.group_validation}
\alias{.group_validation}
\title{Validate all addresses within a data frame of addresses.}
\usage{
.group_validation(df, batch_size = 5)
}
\arguments{
\item{df}{(data.frame) a data frame that must contain the following schema:
Address (character)
City (character)
State (character)
Zip (character)}

\item{batch_size}{(numeric, default 5) specifies the size of a transaction
(maximum 5 for USPS API)}
}
\value{
results (data.frame) a data frame with the same fields as the
  input data frame.
}
\description{
This function takes a dataframe that has been pre-processed with
  the .preprocess_address_data function. It runs through the address data in
  batches to work with the restrictions of the USPS API using
  .batch_query_address until all addresses in the dataframe are verified.
}
