% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_funs.R
\name{.preprocess_address_data}
\alias{.preprocess_address_data}
\title{Reformat address data in preparation for API queries}
\usage{
.preprocess_address_data(df)
}
\arguments{
\item{df}{A tibble that includes columns named "Address", "City", and "Zip".}
}
\value{
The pre-processed tibble with address data.
}
\description{
Preprocess address data before sending to API. This
  includes substituting "#" symbols with empty space, trimming all
  leading and trailing whitespace, and separating state abbreviations
  from the "City" column into their own "State" column.
}
