% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equalization_factors}
\alias{equalization_factors}
\title{Data frame containing Cook County equalization factors}
\format{
A data frame with 40 rows and 2 variables:
\describe{
  \item{year}{The year for the equalization factor}
  \item{equalization_factor}{The equalization factor}
}
}
\usage{
equalization_factors
}
\description{
A dataset of Cook County final equalization factors/multipliers. These
multipliers are calculated by the Illinois Department of Revenue in order to
equalize assessment values across Illinois' counties.
}
\note{
Equalization factors in this data frame are final equalization factors,
tentative factors/multipliers are not included.
}
\keyword{datasets}
