% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{model_axe_recipe}
\alias{model_axe_recipe}
\title{Strip environment data from recipe objects to shrink their size}
\usage{
model_axe_recipe(x)
}
\arguments{
\item{x}{A prepped recipe object created by the recipe package.}
}
\value{
Recipe object from \code{x} but with environment and factor level
  data removed.
}
\description{
Recipe objects used by Tidymodels sometimes become bloated due
to quosures and saved factor levels. This function will strip all data from
a recipe object except the data necessary to use \code{\link[recipes]{bake}}.

This saves an enormous amount of disk space for certain complicated recipes
that need to be saved and re-used.
}
\seealso{
Other model_funs: 
\code{\link{model_axe_tune_data}()},
\code{\link{model_lgbm_cap_num_leaves}()},
\code{\link{model_lgbm_load}()},
\code{\link{model_lgbm_save}()},
\code{\link{model_lgbm_update_params}()}
}
\concept{model_funs}
