% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{model_axe_tune_data}
\alias{model_axe_tune_data}
\title{Strip attribute data from CV objects to shrink their size}
\usage{
model_axe_tune_data(x)
}
\arguments{
\item{x}{Iteration object created by the tune package.}
}
\value{
Iteration object from \code{x} but with CV split and attributes
  removed.
}
\description{
Remove unnecessary attribute data from iteration results objects
created by \code{\link[tune]{tune_grid}} and \code{\link[tune]{tune_bayes}}.
This makes the parameter object significantly smaller when saved to disk.
}
\seealso{
Other model_funs: 
\code{\link{model_axe_recipe}()},
\code{\link{model_get_env}()},
\code{\link{model_lgbm_cap_num_leaves}()},
\code{\link{model_lgbm_load}()},
\code{\link{model_lgbm_save}()},
\code{\link{model_lgbm_update_params}()},
\code{\link{model_predict}()},
\code{\link{rm_intermediate}()}
}
\concept{model_funs}
