% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{model_get_env}
\alias{model_get_env}
\title{Get environmental variable or else a specified default value}
\usage{
model_get_env(x, default)
}
\arguments{
\item{x}{The environmental variable to be loaded, specified as a string.}

\item{default}{The default value to load if the environmental variable is
not set.}
}
\value{
Environmental variable specified in \code{x} or the specified default
  if \code{x} is unset.
}
\description{
Load an environmental variable specified in a .Renviron file OR
load a supplied default value. This function is useful for reports and models
that may depend on the implicit existence of env variables.
}
\examples{
model_get_env("R_CV_ENABLE", TRUE)
model_get_env("R_CV_NUM_FOLDS", 5)
}
