% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{model_lgbm_cap_num_leaves}
\alias{model_lgbm_cap_num_leaves}
\title{Cap the num_leaves hyperparameter for LightGBM models}
\usage{
model_lgbm_cap_num_leaves(params)
}
\arguments{
\item{params}{A dataframe of parameters containing both \code{num_leaves} and
\code{tree_depth}.}
}
\value{
A dataframe of parameters with \code{num_leaves} capped at
  \code{2 ^ max_depth - 1}.
}
\description{
LightGBM grows trees leaf-wise rather than depth-wise, meaning
that not all tree levels will have the same number of leaves.

\code{num_leaves} is implicitly capped at \code{2 ^ max_depth} (full depth
tree) in LightGBM, but saved parameter objects can still show higher numbers.

This function explicitly caps \code{num_leaves} in saved parameter
objects by setting \code{num_leaves} equal to \code{2 ^ max_depth - 1}
where appropriate. Note that in parsnip, \code{max_depth} is substituted by
the \code{tree_depth} hyperparameter.
}
\examples{
params <- data.frame(num_leaves = c(60, 600), tree_depth = c(7, 7))
model_lgbm_cap_num_leaves(params)
}
\seealso{
Other model_funs: 
\code{\link{model_axe_recipe}()},
\code{\link{model_axe_tune_data}()},
\code{\link{model_get_env}()},
\code{\link{model_lgbm_load}()},
\code{\link{model_lgbm_save}()},
\code{\link{model_lgbm_update_params}()},
\code{\link{model_predict}()},
\code{\link{rm_intermediate}()}
}
\concept{model_funs}
