% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{model_lgbm_save}
\alias{model_lgbm_save}
\title{Save a LightGBM model to disk}
\usage{
model_lgbm_save(model, zipfile)
}
\arguments{
\item{model}{A parsnip specification containing a fit LightGBM model.}

\item{zipfile}{A path to save zip file to.}
}
\description{
Save a parsnip model fit object with a LightGBM fit to disk. This is a
a workaround to split the LightGBM model and parsnip specification into
separate files and then zip them together. This is necessary because LightGBM
models really do not like being saved with \code{saveRDS()}.
}
\seealso{
Other model_funs: 
\code{\link{model_axe_recipe}()},
\code{\link{model_axe_tune_data}()},
\code{\link{model_get_env}()},
\code{\link{model_lgbm_cap_num_leaves}()},
\code{\link{model_lgbm_load}()},
\code{\link{model_lgbm_update_params}()},
\code{\link{model_predict}()},
\code{\link{rm_intermediate}()}
}
\concept{model_funs}
