% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nbhd_shp}
\alias{nbhd_shp}
\title{Simple features (sf) dataframe of CCAO neighborhoods.}
\format{
A spatial data frame with 845 geometries and 6 variables.
\describe{
  \item{township_name}{Common name of the township}
  \item{township_code}{Two-digit code used to identify the township}
  \item{triad_code}{Single-digit code of the triad the township is in}
  \item{triad_name}{Common name of the triad the township is in}
  \item{nbhd}{Three-digit assessor neighborhood code, zero-padded}
  \item{geometry}{Attribute column containing sf geometries}
}
}
\usage{
nbhd_shp
}
\description{
An \code{sf} spatial dataframe containing geometries, names, and codes for
CCAO neighborhoods. Generated manually on 5/20/2020. Use for thematic maps
only; these boundaries are recovered from old files and may not be accurate.
}
\keyword{datasets}
