% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{postval_model}
\alias{postval_model}
\title{Create and train a post-modeling adjustment model}
\usage{
postval_model(
  data,
  truth,
  estimate,
  class,
  ntile_group_cols,
  ntile_probs,
  ntile_min_sales,
  ntile_min_turnover,
  ntile_max_abs_adj,
  townhome_group_cols,
  townhome_min_sales,
  townhome_min_turnover,
  townhome_max_abs_adj,
  ratio_cap_upper_bound,
  ratio_cap_lower_bound
)
}
\arguments{
\item{data}{A dataframe containing values to be adjusted, sale prices, as
well as the grouping columns listed in the other arguments.}

\item{truth}{Column in data containing true market values. Typically
sale prices.}

\item{estimate}{Column in data containing estimated market values. Typically
model results.}

\item{class}{Column in data containing the property class. Used to filter
out properties that should not be adjusted by this method.}

\item{ntile_group_cols}{Character vector of column names in the data
used to define ntile-based adjustment groups. Typically this is something
like neighborhood, town code, and modeling group.}

\item{ntile_probs}{The breaks to use for creating and assigning grouping
ntiles. Each property will be "assigned" an ntile based on its estimated
value. This ntile will then be used as an additional grouping variable
when making adjustments.}

\item{ntile_min_sales}{The minimum number of sales within an ntile group
necessary to make a percentage adjustment.}

\item{ntile_min_turnover}{The minimum amount of turnover within an ntile
group necessary to make a percentage adjustment.}

\item{ntile_max_abs_adj}{The maximum absolute percentage adjustment to make
within an ntile group.}

\item{townhome_group_cols}{Character vector of column names in the data
used to define groups of identical homes. Should contain characteristic
columns that would not vary for identical units. For example, number of
square feet, number of bedrooms, exterior wall type, etc.}

\item{townhome_min_sales}{The minimum number of sales within a group needed
to apply a percentage adjustment.}

\item{townhome_min_turnover}{The minimum amount of turnover within a group
needed to apply a percentage adjustment. This is to prevent groups with
dozens of properties from being adjusted by just a few sales.}

\item{townhome_max_abs_adj}{The maximum absolute percentage adjustment to
make within a townhome group.}

\item{ratio_cap_upper_bound}{The upper bound for sales ratios. Properties
with ratios higher than this number will have their estimated value set
exactly at the cap.}

\item{ratio_cap_lower_bound}{The lower bound for sales ratios. Properties
with ratios lower than this number will have their estimated value set
exactly at the cap.}
}
\value{
An object with a class of \code{postval_model} which can be used
to adjust unseen data. See \code{\link{predict.postval_model}}.
}
\description{
In addition to first-pass modeling, the CCAO also runs a much
simpler second-pass model. This second model is internally called
"post-modeling", and is responsible for correcting any deficiencies in
the first model's predictions. Specifically, post-modeling will:
\enumerate{
\item Shift distributions of predicted values such that their median sale
ratio (predicted value / actual sale value) moves toward 1. This is done
by applying a percentage multiplier to all properties within a neighborhood,
modeling group, and estimated value quantile. This percentage multiplier
is capped at 40\% and requires a minimum number of sales and minimum
turnover rate within each adjustment group. This is done to prevent broad
over-or under-assessment. Only a small proportion of all properties receive
this adjustment.
\item Alter assessed values for properties with sales in the last two years
if and only if their sale ratio falls outside specific thresholds. If a
property's sale ratio (predicted value / actual sale value) is less than 0.7
or greater than 2, its predicted value will be adjusted such that its ratio
is exactly at the threshold. For example, if a property sold for \$100K in
2019 and our model predicts that it's value is \$65K, it will have a sale
ratio of 0.65, which is below the 0.7 threshold. As a result, the property's
predicted value will be adjusted to $70K.
\item Ensure that perfectly identical properties are identically valued.
For some property classes, such as townhouses, we manually adjust values
such that all identical properties in the same unit or complex receive the
same predicted value. This is accomplished by replacing individual predicted
values with the median predicted value for all properties in the same unit.
The median value is also adjusted using the process in step 1.
}
}
\seealso{
Other valuation_funs: 
\code{\link{predict.postval_model}()},
\code{\link{val_assign_ntile}()},
\code{\link{val_create_ntiles}()},
\code{\link{val_limit_ratios}()},
\code{\link{val_med_pct_adj}()},
\code{\link{val_townhomes_by_group}()}
}
\concept{valuation_funs}
