% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{predict.postval_model}
\alias{predict.postval_model}
\title{Apply post-modeling adjustments to new data}
\usage{
\method{predict}{postval_model}(object, ..., new_data, truth, estimate)
}
\arguments{
\item{object}{A postval_model object created by the
\code{\link{postval_model}} function.}

\item{...}{Not currently used.}

\item{new_data}{A new dataset with a structure identical to the one used to
train the postval model. It must have the same \code{truth} and
\code{estimate} columns as well as all the grouping columns used
to make the initial model.}

\item{truth}{Column in data containing true market values. Typically
sale prices.}

\item{estimate}{Column in data containing estimated market values. Typically
model results.}
}
\value{
A numeric vector of final, adjusted model values.
}
\description{
Apply post-modeling adjustments to new data
}
\seealso{
Other valuation_funs: 
\code{\link{postval_model}()},
\code{\link{val_assign_ntile}()},
\code{\link{val_create_ntiles}()},
\code{\link{val_limit_ratios}()},
\code{\link{val_med_pct_adj}()},
\code{\link{val_townhomes_by_group}()}
}
\concept{valuation_funs}
