% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recp_funs.R
\name{recp_funs}
\alias{recp_funs}
\alias{recp_clean_keep_dict_vars}
\alias{recp_clean_rename}
\alias{recp_clean_recode}
\alias{recp_clean_relocate}
\alias{recp_feat_time}
\alias{recp_feat_char_inds}
\alias{recp_feat_arms_length}
\title{Clean or alter CCAO data using standardized recipes}
\usage{
recp_clean_keep_dict_vars(data, cols_to_rm = NULL)

recp_clean_rename(data, ...)

recp_clean_recode(data)

recp_clean_relocate(data)

recp_feat_time(data, origin_date = "1997-01-01")

recp_feat_char_inds(data)

recp_feat_arms_length(data)
}
\arguments{
\item{data}{A data frame containing data extracted from CCAO SQL tables.
Expects either SQL or "standardized" column names as listed in
\code{\link{vars_dict}}.}

\item{cols_to_rm}{Additional columns to remove when discarding all
columns not present in \code{\link{vars_dict}}.}

\item{...}{Arguments passed to \code{\link{vars_rename}}.}

\item{origin_date}{Origin date for calculated time variables. For example,
\code{time_sale_day} will be equal to the number of days between
\code{origin_date} and date of sale.}
}
\value{
A data frame after the recipes steps are applied.
}
\description{
Data extracted from CCAO SQL tables often requires cleaning and
transformation before modeling. These functions standardized such actions
into "recipes", similar to the R recipes package. They should be applied
in the order listed below:
}
\section{Functions}{
\itemize{
\item \code{recp_clean_keep_dict_vars}: Remove all variables not present in
\code{\link{vars_dict}} and any additional variables listed in
\code{cols_to_rm}.

\item \code{recp_clean_rename}: Rename variable names from SQL to the "standard" names
contained in \code{\link{vars_dict}}.

\item \code{recp_clean_recode}: Recode variables to their expected types and clean up
other variables like date and age.

\item \code{recp_clean_relocate}: Reorder variables into consistent groups by
standardized name (meta, char, etc.).

\item \code{recp_feat_time}: Create time fixed effect variables for sales using
1997 as a base year.

\item \code{recp_feat_char_inds}: Create indicator variables on characteristics for
use in modeling.

\item \code{recp_feat_arms_length}: Represents user defined exclusions due to data issues.
Observations with ind_arms_length = 1 have "pure market" qualities and
good data.
}}

\concept{recp_funs}
