% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{rm_intermediate}
\alias{rm_intermediate}
\title{Remove intermediate objects generated by model creation}
\usage{
rm_intermediate(x, keep = NULL)
}
\arguments{
\item{x}{Prefix of the objects to be removed from the global environment.}

\item{keep}{A character vector of objects with the prefix specified in
\code{x} that should NOT be removed.}
}
\description{
Recover limited memory resources consumed by large objects
during modeling. This function will remove anything with the prefix passed
to \code{x}, then will run garbage collection to free memory. Will not remove
any objects with "final" in their name.
}
\examples{
rm_intermediate("lgbm", keep = "lgbm_params")
}
\seealso{
Other model_funs: 
\code{\link{model_axe_recipe}()},
\code{\link{model_axe_tune_data}()},
\code{\link{model_lgbm_cap_num_leaves}()},
\code{\link{model_lgbm_load}()},
\code{\link{model_lgbm_save}()},
\code{\link{model_lgbm_update_params}()}
}
\concept{model_funs}
