% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{town_shp}
\alias{town_shp}
\title{Simple features (sf) dataframe of Cook County townships.}
\format{
A spatial data frame with 38 geometries and 3 variables.
\describe{
  \item{township_name}{Common name of the township}
  \item{township_code}{Two-digit code used to identify the township}
  \item{triad_code}{Single-digit code of the triad the township is in}
  \item{triad_name}{Common name of the triad the township is in}
  \item{geometry}{Attribute column containing sf geometries}
}
}
\usage{
town_shp
}
\description{
An \code{sf} spatial dataframe containing geometries, names, and codes for
Cook County political/administrative townships. Use for thematic
mapping only.
}
\keyword{datasets}
