% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{val_assign_ntile}
\alias{val_assign_ntile}
\title{Assign an ntile to a numeric value}
\usage{
val_assign_ntile(x, ntiles)
}
\arguments{
\item{x}{A numeric vector to assign ntiles to.}

\item{ntiles}{A quantile or output from \code{\link{val_create_ntiles}}.}
}
\value{
A character vector with the assigned ntile range
  for each value of \code{x}.
}
\description{
Assign a numeric value its respective ntile or quantile, given
a pre-made set. This function is used in conjunction with
\code{\link{val_create_ntiles}}.
}
\examples{
x <- 1:100
cuts <- list(stats::quantile(x, probs = c(0.25, 0.75)))

ccao:::val_assign_ntile(x, cuts)

}
