% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{val_create_ntiles}
\alias{val_create_ntiles}
\title{Create a list of ntiles based on an input distribution.}
\usage{
val_create_ntiles(x, probs, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector from which to determine ntiles.}

\item{probs}{Quantile probabilities or splits.}

\item{na.rm}{Whether or not to remove NAs from inputs.}
}
\value{
A list containing the ntile breakpoints, bounded by -Inf and Inf.
}
\description{
Create ntiles based on a numeric distribution. Min and max are
replaced with -Inf and Inf so that new values outside the initial range can
be used. This function is used in conjunction with
\code{\link{val_assign_ntile}}.
}
\examples{
x <- 1:1000
ccao:::val_create_ntiles(x, probs = c(0.25, 0.50, 0.75))

estimates <- c(30000, 50000, 31000, 10000, 200000, 30000, 20000)
ccao:::val_create_ntiles(estimates, probs = c(0.25, 0.50, 0.75))
}
