% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{val_med_pct_adj}
\alias{val_med_pct_adj}
\title{Calculate median percentage adjustments}
\usage{
val_med_pct_adj(truth, estimate, min_n, max_abs_adj, na.rm = TRUE)
}
\arguments{
\item{truth}{Vector of true market values. Typically sale prices.}

\item{estimate}{Vector of estimated market values. Typically model results.}

\item{min_n}{Minimum number of true values needed to make an adjustment. If
the number of true values is less than \code{min_n}, then the function will
return NA.}

\item{max_abs_adj}{The maximum absolute percentage adjustment that can be
returned.}

\item{na.rm}{Whether or not to remove NAs from inputs.}
}
\value{
A single percentage adjustment. Can be applied to estimated values to
  shift them toward the true values.
}
\description{
In cases where models create a biased output or have a skewed
error structure, this function can be used to calculate a fixed percentage
adjustment that will shift the distribution of errors toward the "true"
values.

This should not be used lightly. In the case of adjusting modeled values
toward "true" sale prices, it should only be used with high confidence that
those sales are representative of the overall market.

In practice, the primary use of this function is to adjust ratios at the very
low and very high end of the market.
}
\examples{
sales <- c(20000, 15000, 30000, NA, 55000, 40000, NA)
estimates <- c(30000, 20000, 31000, 10000, 100000, 30000, 20000)
val_med_pct_adj(sales, estimates, min_n = 3, max_abs_adj = 0.4)
}
\seealso{
Other valuation_funs: 
\code{\link{postval_model}()},
\code{\link{predict.postval_model}()},
\code{\link{val_assign_ntile}()},
\code{\link{val_create_ntiles}()},
\code{\link{val_limit_ratios}()},
\code{\link{val_townhomes_by_group}()}
}
\concept{valuation_funs}
