% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{val_round_fmv}
\alias{val_round_fmv}
\title{Round model values to flat amounts}
\usage{
val_round_fmv(x, breaks = 1e+05, round_to = c(5000, 10000), type = "floor")
}
\arguments{
\item{x}{A numeric vector of market or assessed values. All values must be
greater than or equal to 0.}

\item{breaks}{A numeric vector of breaks used to split \code{x} into ranges
that receive different rounding amounts. The default splits \code{x} at
$100K, with values above $100K rounded down to the nearest $10K, and values
below $100K rounded down to the nearest $5K.}

\item{round_to}{A numeric vector of values to round to. Must be 1 longer than
\code{breaks}. Default rounds down to the nearest $5K and $10K.}

\item{type}{A character value indicating the type of rounding to use.
Options are \code{"floor"}, \code{"normal"}, and \code{"ceiling"}. Default
is \code{"floor"}.}
}
\value{
A numeric vector of rounded values.
}
\description{
Estimated values produced by the CCAO are sometimes rounded to
  a nearby flat amount e.g. nearest $5K, nearest $100, etc. This is done
  to indicate that model values are estimates, rather than precise
  amounts.

  The function supports applying different rounding amounts to different
  ranges of market values. For example, estimated property values below
  $200K might be rounded to the nearest $1,000, while estimated property
  values below $1M might be rounded to the nearest $5,000.
}
