% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{val_townhomes_by_group}
\alias{val_townhomes_by_group}
\title{Adjust identical properties by group}
\usage{
val_townhomes_by_group(
  data,
  truth,
  estimate,
  class,
  townhome_group_cols,
  townhome_min_sales,
  townhome_min_turnover,
  townhome_max_abs_adj
)
}
\arguments{
\item{data}{A dataframe containing property values to be adjusted.}

\item{truth}{Column in data containing true market values. Typically
sale prices.}

\item{estimate}{Column in data containing estimated market values. Typically
model results.}

\item{class}{Column in data containing the property class. Used to filter
out properties that should not be adjusted by this method.}

\item{townhome_group_cols}{Character vector of column names in the data
used to define groups of identical homes. Should contain characteristic
columns that would not vary for identical units. For example, number of
square feet, number of bedrooms, exterior wall type, etc.}

\item{townhome_min_sales}{The minimum number of sales within a group needed
to apply a percentage adjustment.}

\item{townhome_min_turnover}{The minimum amount of turnover within a group
needed to apply a percentage adjustment. This is to prevent groups with
dozens of properties from being adjusted by just a few sales.}

\item{townhome_max_abs_adj}{The maximum absolute percentage adjustment to
make within a townhome group.}
}
\value{
A summarized dataframe with the adjustment and final value of each
  property group.
}
\description{
Within the same development, townhomes with the same
characteristics (nearly identical) should have the same value. This function
groups such properties and then sets their estimated value to the median of
the group. It will also apply a median adjustment if enough sales are
available within the group.
}
\seealso{
Other valuation_funs: 
\code{\link{postval_model}()},
\code{\link{predict.postval_model}()},
\code{\link{val_assign_ntile}()},
\code{\link{val_create_ntiles}()},
\code{\link{val_limit_ratios}()},
\code{\link{val_med_pct_adj}()}
}
\concept{valuation_funs}
