% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_funs.R
\name{validate_addresses}
\alias{validate_addresses}
\title{Verify and reformat address information with the USPS API.}
\usage{
validate_addresses(address, city, zip, batch_size = 5)
}
\arguments{
\item{address}{(character) string or vector of addresses ({number}
{streetname} {unit if applicable}). Either 1 long or the same
length as \code{city} and \code{zip}.}

\item{city}{(character) string or vector of city and state
associated with an address ({city} {state}). Either 1 long
or the same length as \code{address} and \code{zip}.}

\item{zip}{(character) string or vector of 5-digit zip codes
associated with an address. Either 1 long or the same length
as \code{address} and \code{city}.}

\item{batch_size}{(default 5) specifies the number of queries to be
sent to API at a time (maximum 5).}
}
\value{
validated_addresses (character) string or vector of verified
  addresses formatted the following way:
    "{address}, {city}, {state}, {zip}".
  Throws error if input data types are incorrect or if length
  conditions of input vectors are not met.
}
\description{
Bulk verify address information using street address,
  state, city, and zip code data and return a single vector of
  complete, verified, comma-separated addresses.

  Note that you must create an API key with USPS and set it as
  environmental variable called "USPS_API_KEY" on your local
  machine for this function to work. This can be done by typing
  \code{Sys.setenv("USPS_API_KEY" = "XXXXXXXXXX")} in the RStudio
  console.
}
\concept{address_funs}
