% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_funs.R
\name{vars_check_class}
\alias{vars_check_class}
\title{Check if a property class falls within its expected square footage and age
boundaries}
\usage{
vars_check_class(age, sqft, class)
}
\arguments{
\item{age}{Integer or numeric vector of ages of properties. Either 1 long
or the same length as \code{sqft} and \code{class}.}

\item{sqft}{Integer or numeric vector of the square footage of properties.
Either 1 long or the same length as \code{age} and \code{class}.}

\item{class}{String or character vector of class codes. Either 1 long or
the same length as \code{sqft} and \code{age}.}
}
\value{
A logical vector indicating that the specified class falls within
  the parameters specified by \code{\link{class_dict}}. Throws error if input
  data types are incorrect or if length conditions of input vectors
  are not met.
}
\description{
Check property characteristics against class definitions as defined # nolint
\href{https://datascience.cookcountyassessor.com/wiki/data/class-definitions.pdf}{here}. # nolint
}
\examples{
vars_check_class(50, 800, "202")
vars_check_class(c(50, 80), c(800, 1000), c("202", "203"))
vars_check_class(c(50, 80), 1000, "210")
vars_check_class(50, c(800, 2000), "202")
vars_check_class(50, 1000, c("202", "203"))
}
\seealso{
Other vars_funs: 
\code{\link{vars_recode}()},
\code{\link{vars_rename}()}
}
\concept{vars_funs}
