% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_funs.R
\name{vars_rename}
\alias{vars_rename}
\title{Translate variable names from different CCAO data sources}
\usage{
vars_rename(data, names_from = "sql", names_to = "standard")
}
\arguments{
\item{data}{A data frame or tibble with columns to be renamed.}

\item{names_from}{The source/name type of data. See description}

\item{names_to}{The target names. See description}
}
\value{
The input data frame with columns renamed.
}
\description{
Bulk rename columns from one type of CCAO to another. For
example, rename all columns pulled from SQL to their standard names used
in modeling. Or, rename all standard modeling names to "pretty" names for
publication. Function will overwrite names it finds in 
\code{\link{vars_dict}}, all other names in the data will remain unchanged.

Options for \code{names_from} and \code{names_to} are: \code{"sql"} (
with names like TAX_YEAR, GAR1_SIZE); \code{"addchars"} (with names like
QU_AGE, QU_GARAGE_SIZE); \code{"standard"} (with names like meta_tax_year,
meta_gar1_size); and \code{"pretty"} (with names like Year, Garage 1 Size)
}
\examples{

# Rename column names from SQL
vars_rename(chars_sample_universe)
vars_rename(chars_sample_universe, names_to = "pretty")

# No renames will occur since no column names here are from SQL
vars_rename(class_dict)
}
\seealso{
Other vars_funs: 
\code{\link{vars_check_class}()},
\code{\link{vars_recode}()}
}
\concept{vars_funs}
