% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccao_funs.R
\name{ccao_download_input_data}
\alias{ccao_download_input_data}
\title{Download one or more DVC-tracked input datasets for a given model run}
\usage{
ccao_download_input_data(
  model_run,
  files,
  s3_staging_dir = "s3://ccao-athena-results-us-east-1"
)
}
\arguments{
\item{model_run}{Character `run_id` found in `model.metadata`.}

\item{files}{Character vector of file keys to download. Valid keys are:
`assessment`, `complex_id`, `land_nbhd`, `land_site`, `training`, `char`,
`hie`, `condo_strata`.}

\item{s3_staging_dir}{S3 path to use for Athena query results. This is
mainly used to pass checks in github.}
}
\value{
If `files` has length 1, returns a single data frame (as returned by
`arrow::read_parquet()`). If `files` has length > 1, returns a named list of
data frames with names equal to `files`.
}
\description{
Downloads one or more model input datasets referenced by a
`run_id` in `model.metadata`.

Folder layout depends on the model year and assessment group:
- For `assessment_year <= 2024`, the legacy DVC layout is used.
- For `assessment_year >= 2025`, the path includes `model-res-avm` or
  `model-condo-avm` depending on `assessment_group`.
}
