% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccao_funs.R
\name{ccao_download_model_input_data}
\alias{ccao_download_model_input_data}
\title{Download one or more DVC-tracked input datasets for a given model run}
\usage{
ccao_download_model_input_data(
  run_id,
  file_keys,
  s3_staging_dir = "s3://ccao-athena-results-us-east-1"
)
}
\arguments{
\item{run_id}{Character `run_id` found in `model.metadata`.}

\item{file_keys}{Character vector of file keys to download. Valid keys are:
`assessment`, `complex_id`, `land_nbhd_rate`, `land_site_rate`,
`training`, `char`, `hie`, `condo_strata`.}

\item{s3_staging_dir}{S3 path to use for Athena query results. This is
mainly used to pass checks in github.}
}
\value{
If `file_keys` is a single key, returns a single DataFrame.
If `file_keys` is a list, returns a dict of DataFrames keyed by file key.
}
\description{
Downloads one or more model input datasets referenced by a
`run_id` in `model.metadata`.
}
