% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccao_funs.R
\name{ccao_generate_id}
\alias{ccao_generate_id}
\title{Generate a memorable ID string based on CCAO employee names}
\usage{
ccao_generate_id(n = 1L, suffix = as.character(Sys.Date()))
}
\arguments{
\item{n}{Integer of the number of IDs to generate. Default 1.}

\item{suffix}{Suffix to append to the end of each ID, uses the system date
by default. Set to NULL for no suffix or a vector of length n for multiple
different suffixes.}
}
\value{
An n-long character vector of memorable IDs.
}
\description{
Memorable IDs are useful for referencing objects with arbitrary
versions. This function uses the
\href{https://github.com/moby/moby/blob/master/pkg/namesgenerator/names-generator.go}{Docker name generator code}
to create random memorable IDs using Data Department employee, interns, and
fellow names.
}
\examples{
# Generate a single ID
ccao_generate_id()

# Multiple IDs
ccao_generate_id(10)

# IDs with no suffix
ccao_generate_id(3, NULL)

# Different suffix for each ID
ccao_generate_id(3, c("type1", "type2", "type3"))
}
