% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chars_sample_athena}
\alias{chars_sample_athena}
\title{Sample dataset from the vw_pin_universe Athena SQL view}
\format{
An object of class \code{data.frame} with 50 rows and 35 columns.
}
\source{
This data was extracted from Athena manually on 2022-01-13.
}
\usage{
chars_sample_athena
}
\description{
A dataset containing a small subsample of rows from the
default.vw_pin_universe Athena view.
}
\keyword{datasets}
