% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_funs.R
\name{.batch_query_address}
\alias{.batch_query_address}
\title{Perform a single batched address query to the USPS API}
\usage{
.batch_query_address(address, city, state, zip, batch_size = 5, api_key)
}
\arguments{
\item{address}{Character string or vector of addresses ({number}
{streetname} {unit if applicable}). Either 1 long or the same length
as \code{city}, \code{zip}, and \code{state}.}

\item{city}{Character string or vector of city and state associated with
an address ({city} {state}). Either 1 long or the same length as
\code{address}, \code{zip}, and \code{state}.}

\item{state}{Character string or vector of 2-character state abbreviations
associated with an address. Either 1 long or the same length as
\code{address}, \code{city}, and \code{zip}.}

\item{zip}{Character string or vector of 5-digit zip codes associated with
an address. Either 1 long or the same length as \code{address},
\code{city}, and \code{state}.}

\item{batch_size}{Default 5. Specifies the number of queries to be sent
to API at a time (maximum 5).}

\item{api_key}{Specify a required USPS API key.}
}
\value{
A tibble with verified address information from the API for one batch
  of at most 5 addresses. If an input address is found to be nonexistent,
  it will not show up in the returned tibble.
}
\description{
Generate 1-5 queries from the USPS API
  (max transaction size = 5)
}
