% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_funs.R
\name{.group_validation}
\alias{.group_validation}
\title{Validate all addresses within a tibble of addresses}
\usage{
.group_validation(df, batch_size = 5, api_key = "USPS_API_KEY")
}
\arguments{
\item{df}{A tibble that must contain the following schema:
Address (character)
City (character)
State (character)
Zip (character)}

\item{batch_size}{Default 5. Specifies the size of a transaction
(maximum 5 for USPS API)}

\item{api_key}{Default "USPS_API_KEY". Specifies the name of an environment
variable in which the user's API key is saved. Will not run if the API key
is an empty string, or if it is invalid.}
}
\value{
A tibble with the same fields as the input tibble.
}
\description{
This function takes a tibble that has been pre-processed with
  the .preprocess_address_data function. It runs through the address data in
  batches to work with the restrictions of the USPS API using
  .batch_query_address until all addresses in the tibble are verified.
}
