% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{model_lgbm_load}
\alias{model_lgbm_load}
\title{Load a LightGBM model from disk}
\usage{
model_lgbm_load(zipfile)
}
\arguments{
\item{zipfile}{A path to the zip file containing the saved model objects.}
}
\value{
A parsnip model fit with LightGBM model.
}
\description{
This is the paired function to \code{model_lgbm_save()}. It loads a zip file
containing a parsnip specification and LightGBM model, then recombines them
into a single object.
}
