% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{model_predict}
\alias{model_predict}
\title{Predict using a trained model and recipe}
\usage{
model_predict(spec, recipe, data, exp = TRUE)
}
\arguments{
\item{spec}{A parsnip model specification object. Must be trained.}

\item{recipe}{A prepped (trained) recipe object. Must be trained.}

\item{data}{New data to get predictions from. Will be preprocessed by the
specified \code{recipe}.}

\item{exp}{Exponentiate the returned prediction (assumed to be log). Default
TRUE.}
}
\value{
A vector of predictions from the model given the data and recipe
  specified.
}
\description{
Simple helper function to return predictions from a new data set
given a parsnip specification and recipe. Will exponentiate predictions by
default.
}
