% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_funs.R
\name{validate_addresses}
\alias{validate_addresses}
\title{Verify and reformat address information with the USPS API}
\usage{
validate_addresses(address, city, state, zip, batch_size = 5, api_key)
}
\arguments{
\item{address}{Character string or vector of addresses ({number}
{streetname} {unit if applicable}). Either 1 long or the same length
as \code{city}, \code{zip}, and \code{state}.}

\item{city}{Character string or vector of city and state associated with
an address ({city} {state}). Either 1 long or the same length as
\code{address}, \code{zip}, and \code{state}.}

\item{state}{Character string or vector of 2-character state abbreviations
associated with an address. Either 1 long or the same length as
\code{address}, \code{city}, and \code{zip}.}

\item{zip}{Character string or vector of 5-digit zip codes associated with
an address. Either 1 long or the same length as \code{address},
\code{city}, and \code{state}.}

\item{batch_size}{Default 5. Specifies the number of queries to be
sent to API at a time (maximum 5).}

\item{api_key}{Specify a required USPS API key.}
}
\value{
A string or vector of verified addresses formatted the following way:
    "{address}, {city}, {state}, {zip}".
  Throws error if input data types are incorrect or if length
  conditions of input vectors are not met.
}
\description{
Bulk verify address information using street address,
  state, city, and zip code data and return a single vector of
  complete, verified, comma-separated addresses.

  Note that a USPS API key is required to use this function.
}
\examples{
address <- "3726 N WILTON"
city <- "CHICAGO  "
state <- "IL"
zip <- "60613"
\dontrun{
validate_addresses(address, city, state, zip, api_key = "API_KEY")
}
}
\concept{address_funs}
