% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vars_dict}
\alias{vars_dict}
\title{Data dictionary for CCAO variables}
\format{
A data frame with 175 rows and 14 variables:
\describe{
  \item{var_name_sql}{Column name of variable as stored in CCAO SQL servers}
  \item{var_name_addchars}{Column name of variable when stored in the
  ADDCHARS SQL table}
  \item{var_name_socrata}{Column name used for the Cook County Data Portal}
  \item{var_name_standard}{Column name used for flat files, open data, and
  modeling}
  \item{var_name_pretty}{Human-readable column name used for publication}
  \item{var_type}{Variable type: meta = identifying information; char =
  characteristics retrieved from SQL; time = calculated time variable; ind =
  calculated logical/indicator variable; econ = calculated economic
  information; geo = attached/calculated geographic information}
  \item{var_data_type}{R data type variable values should be stored as}
  \item{var_is_published}{Logical value indicating whether to publish in
  open data}
  \item{var_is_clustered}{Logical value indicating whether to use variable
  in comparables clustering algorithm}
  \item{var_is_predictor}{Logical value indicating whether to use variable
  in modeling on the right-hand side. Left-hand side is always sale price}
  \item{var_code}{Factor value for categorical variable. These are the values
  stored in the AS/400}
  \item{var_value}{Human-readable translation of factor value}
  \item{var_value_short}{Human-readable translation of factor value, but as
  short as possible}
  \item{var_notes}{Field descriptions and caveats}
}
}
\source{
This dictionary was manually created from paper forms as a
  translation of numeric variables. char_value_short is the equivalent of
  what is used on the AS/400 property info screens
}
\usage{
vars_dict
}
\description{
A crosswalk of CCAO variable names used in SQL, modeling, and open data. Also
includes a translation of numeric character codes to their human-readable
value (ROOF_CNST = 1 becomes ROOF_CNST = Shingle/Asphalt).
}
\keyword{datasets}
