% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_funs.R
\name{vars_rename}
\alias{vars_rename}
\title{Bulk rename variables from CCAO SQL to standardized or pretty names
and visa versa}
\usage{
vars_rename(data, names_from = "sql", names_to = "standard", type = "inplace")
}
\arguments{
\item{data}{A data frame or tibble with columns to be renamed.}

\item{names_from}{The source/name type of data. See description}

\item{names_to}{The target names. See description}

\item{type}{Output type. Either \code{"inplace"}, which renames the input
data frame, or \code{"vector"}, which returns a named character vector with
the construction new_col_name = old_col_name.}
}
\value{
The input data frame with columns renamed.
}
\description{
Bulk rename columns from one type of CCAO to another. For
example, rename all columns pulled from SQL to their standard names used
in modeling. Or, rename all standard modeling names to "pretty" names for
publication. Function will overwrite names it finds in
\code{\link{vars_dict}}, all other names in the data will remain unchanged.

Options for \code{names_from} and \code{names_to} are:
\itemize{
\item \code{"sql"} (with names like TAX_YEAR, GAR1_SIZE)
\item \code{"addchars"} (with names like QU_AGE, QU_GARAGE_SIZE)
\item \code{"socrata"} (with names like tax_year, gar1_size)
\item \code{"standard"} (with names like meta_tax_year, meta_gar1_size)
\item \code{"pretty"} (with names like Year, Garage 1 Size)
}
}
\examples{

# Rename column names from SQL
sample_data <- chars_sample_universe[1:5, 18:27]

vars_rename(sample_data)
vars_rename(sample_data, names_to = "pretty")

# No renames will occur since no column names here are from SQL
vars_rename(class_dict[1:5, 1:5])
}
\seealso{
Other vars_funs: 
\code{\link{vars_check_class}()},
\code{\link{vars_recode}()}
}
\concept{vars_funs}
